+++ hs-bindgen-runtime/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 0.1.0 -- YYYY-mm-dd
 
-* First version. Released on an unsuspecting world.
+* NON-BREAKING: add new `safeCastFunPtr` function to `HsBindgen.Runtime.Prelude`